Move python setup utilities into it's own file
This was bloating the `extensions.ts` file, and we needed it next PR for the utils checking
