+++ vscode-extension/src/extension.ts
@@ -15,17 +15,13 @@ import {
   sanitizeFileName,
   getTodayDateString,
   LASTMILE_BASE_URI,
-  getPythonPath,
   isSupportedConfigExtension,
   SUPPORTED_FILE_EXTENSIONS,
-  isPythonVersionAtLeast310,
-  showGuideForPythonInstallation,
   setupEnvironmentVariables,
-  savePythonInterpreterToCache,
 } from "./util";
+import { initialize, getPythonPath } from "./utilities/pythonSetupUtils";
 import { AIConfigEditorProvider } from "./aiConfigEditor";
 import { AIConfigEditorManager } from "./aiConfigEditorManager";
-import { PythonExtension } from "@vscode/python-extension";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -424,302 +420,6 @@ async function handleCustomModelRegistryUpdate(
   );
 }
 
-/**
- * Installs the dependencies required for the AIConfig extension to work
- */
-async function installDependencies(
-  context: vscode.ExtensionContext,
-  outputChannel: vscode.LogOutputChannel
-) {
-  vscode.window.withProgress(
-    {
-      location: vscode.ProgressLocation.Notification,
-      title: "Initializing AIConfig Extension",
-      cancellable: false,
-    },
-    async (progress, cancellationToken) => {
-      outputChannel.appendLine("Initializing extension");
-
-      outputChannel.appendLine("1. Making sure Python is installed");
-      progress.report({
-        increment: 0,
-        message: "Checking Python installation",
-      });
-
-      const isPythonInstalled = await checkPython();
-      if (!isPythonInstalled) {
-        outputChannel.appendLine("Python is not installed");
-        return;
-      }
-      await savePythonInterpreterToCache();
-
-      outputChannel.append(" -- SUCCESS");
-      outputChannel.appendLine("2. Making sure pip is installed");
-
-      progress.report({
-        increment: 20,
-        message: "Checking pip installation",
-      });
-
-      const isPipInstalled = await checkPip();
-      if (!isPipInstalled) {
-        outputChannel.appendLine("Pip is not installed");
-        return;
-      }
-
-      outputChannel.append(" -- SUCCESS");
-      outputChannel.appendLine("3. Checking dependencies to install or update");
-
-      progress.report({
-        increment: 20,
-        message: "Checking dependencies to install or update",
-      });
-
-      // Check if requirements need to be installed
-      const requirementsInstalled = await checkRequirements(
-        context,
-        outputChannel
-      );
-      if (requirementsInstalled) {
-        outputChannel.append(" -- SUCCESS");
-        // Requirements are already installed
-        progress.report({
-          increment: 40,
-          message: "Dependencies are already installed. Let's go!",
-        });
-      } else {
-        outputChannel.appendLine("Update required. Installing dependencies");
-
-        // Install or update requirements
-        progress.report({
-          increment: 20,
-          message: "Installing dependencies",
-        });
-
-        const installationResult = await installRequirements(
-          context,
-          progress,
-          cancellationToken,
-          outputChannel
-        );
-        if (!installationResult) {
-          // The installation encountered issues -- the installRequirements function will have already shown an error message
-          outputChannel.error(
-            "Failed to install dependencies. Please try again."
-          );
-          return;
-        } else {
-          // Installation was successful
-          progress.report({
-            increment: 40,
-            message: "Dependencies installed. Let's go!",
-          });
-
-          outputChannel.append(" -- SUCCESS");
-        }
-      }
-    }
-  );
-}
-
-/**
- * Installs the packages in the requirements.txt file needed for the AIConfig extension to work.
- */
-async function installRequirements(
-  context: vscode.ExtensionContext,
-  progress: vscode.Progress<{
-    message?: string | undefined;
-    increment?: number | undefined;
-  }>,
-  cancellationToken: vscode.CancellationToken,
-  outputChannel: vscode.LogOutputChannel
-) {
-  const extensionPath = context.extensionPath;
-  const requirementsPath = path.join(
-    extensionPath,
-    "python",
-    "requirements.txt"
-  );
-  const pythonPath = await getPythonPath();
-
-  return new Promise((resolve, _reject) => {
-    const pipInstall = spawn(pythonPath, [
-      "-m",
-      "pip",
-      "install",
-      "-r",
-      requirementsPath,
-      "--upgrade",
-    ]);
-
-    pipInstall.stdout.on("data", (data) => {
-      progress.report({
-        message: `Installing dependencies: ${data}`,
-      });
-      outputChannel.info(`pip install: ${data}`);
-      console.log(`pip install stdout: ${data}`);
-    });
-
-    pipInstall.stderr.on("data", (data) => {
-      outputChannel.error(`pip install: ${data}`);
-      console.error(`pip install stderr: ${data}`);
-    });
-
-    pipInstall.on("close", (code) => {
-      if (code !== 0) {
-        console.log(`pip install process exited with code ${code}`);
-        vscode.window
-          .showErrorMessage(
-            `Failed to install dependencies. Pip exited with code ${code}. Please try again later`,
-            ...["Select Interpreter", "Retry Install Dependencies"]
-          )
-          .then((selection) => {
-            if (selection === "Retry Install Dependencies") {
-              installRequirements(
-                context,
-                progress,
-                cancellationToken,
-                outputChannel
-              );
-            } else if (selection === "Select Interpreter") {
-              vscode.commands.executeCommand(COMMANDS.INIT);
-            }
-          });
-        resolve(false);
-      } else {
-        resolve(true);
-      }
-    });
-  });
-}
-
-/**
- * Runs the check_requirements.py script to check if any requirements need to be updated or installed.
- */
-async function checkRequirements(
-  context: vscode.ExtensionContext,
-  outputChannel: vscode.LogOutputChannel
-) {
-  const extensionPath = context.extensionPath;
-  const checkRequirementsScriptPath = path.join(
-    extensionPath,
-    "python",
-    "src",
-    "check_requirements.py"
-  );
-
-  const requirementsPath = path.join(
-    extensionPath,
-    "python",
-    "requirements.txt"
-  );
-
-  const pythonPath = await getPythonPath();
-
-  return new Promise((resolve, reject) => {
-    let checkRequirements = spawn(pythonPath, [
-      checkRequirementsScriptPath,
-      "--requirements_path",
-      requirementsPath,
-    ]);
-
-    checkRequirements.stdout.on("data", (data) => {
-      outputChannel.info(`check_requirements: ${data}`);
-    });
-
-    checkRequirements.on("close", (code) => {
-      if (code !== 0) {
-        // Need to install requirements
-        const msg = "Requirements are not installed or out of date";
-        outputChannel.warn(`check_requirements: ${msg}`);
-        console.log(msg);
-        resolve(false);
-      } else {
-        // Requirements are installed
-        const msg = "Requirements are already installed";
-        outputChannel.info(`check_requirements: ${msg}`);
-
-        console.log(msg);
-        resolve(true);
-      }
-    });
-
-    checkRequirements.on("error", (err) => {
-      outputChannel.error(err);
-      reject(err);
-    });
-  });
-}
-
-/**
- * Checks if Python is installed on the system, and if not, prompts the user to install it.
- */
-async function checkPython() {
-  const pythonPath = await getPythonPath();
-  return new Promise((resolve, _reject) => {
-    exec(pythonPath + " --version", (error, stdout, stderr) => {
-      if (error) {
-        console.error("Python was not found, can't install requirements");
-        console.error("retrieved python path: " + pythonPath);
-        console.error("error: " + error);
-
-        // Guide for installation
-        showGuideForPythonInstallation(
-          "Specified Python Interpreter is not valid"
-        );
-        resolve(false);
-      } else if (!isPythonVersionAtLeast310(pythonPath)) {
-        showGuideForPythonInstallation(
-          "Python version is not 3.10 or higher. Please upgrade to Python 3.10 or higher."
-        );
-        console.error(
-          "Python version is not 3.10 or higher. Please upgrade to Python 3.10 or higher."
-        );
-        resolve(false);
-        // show guide for installation
-      } else {
-        resolve(true);
-        console.log("Python is installed");
-      }
-    });
-  });
-}
-
-/**
- * Checks if pip is installed on the system, and if not, prompts the user to install it.
- */
-async function checkPip() {
-  const pythonPath = await getPythonPath();
-  return new Promise((resolve, _reject) => {
-    // when calling pip using `python -m`, no need to worry about pip vs pip3.
-    // You're directly specifying which Python environment's pip to use.
-    exec(pythonPath + " -m pip --version", (error, stdout, stderr) => {
-      if (error) {
-        console.log("pip is not found");
-        // Guide for installation
-        vscode.window
-          .showErrorMessage(
-            "pip is not installed, but is needed for AIConfig installation",
-            ...["Install pip", "Retry"]
-          )
-          .then((selection) => {
-            if (selection === "Install pip") {
-              vscode.env.openExternal(
-                vscode.Uri.parse("https://pip.pypa.io/en/stable/installation/")
-              );
-            } else if (selection === "Retry") {
-              vscode.commands.executeCommand(COMMANDS.INIT);
-            }
-          });
-        resolve(false);
-      } else {
-        console.log("pip is installed");
-        resolve(true);
-      }
-    });
-  });
-}
-
 async function shareAIConfig(
   context: vscode.ExtensionContext,
   aiconfigEditorManager: AIConfigEditorManager
@@ -816,19 +516,4 @@ async function shareAIConfig(
         }
       });
   }
-}
-
-/**
- * Selects a Python interpreter and initiates the dependency installation flow.
- */
-async function initialize(
-  context: vscode.ExtensionContext,
-  outputChannel: vscode.LogOutputChannel
-) {
-  // Make sure Python API is activated
-  const pythonApi: PythonExtension = await PythonExtension.api();
-
-  await vscode.commands.executeCommand("python.setInterpreter");
-
-  installDependencies(context, outputChannel);
 }
